% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{euclidean function}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
the greatest common divisor of two numbers
}
\description{
algorithm to find the greatest common divisor of two numbers
}
\details{
the greatest common divisor of two numbers
}
\examples{
euclidean(123612, 13892347912)
euclidean(123612, 13892347912)
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean algorithm}
}
