\name{Lab3-package}
\alias{Lab3-package}
\alias{Lab3}
\docType{package}
\title{
\packageTitle{Lab3}
}
\description{
\packageDescription{Lab3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Lab3}
\packageIndices{Lab3}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}
\author{
\packageAuthor{Lab3}

Maintainer: \packageMaintainer{Lab3}
}

\keyword{ package }

