% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Nodes and the weight between each two nodes.
a data.frame with three variables (v1, v2 and w)}
\format{
A data frame with 3 variables:
\describe{
  \item{v1}{Source Node}
  \item{v2}{Destination Node}
  \item{w}{The wirght between the source and destination Node}
}
}
\usage{
wiki_graph
}
\description{
Nodes and the weight between each two nodes.
a data.frame with three variables (v1, v2 and w)
}
\references{
\url{https://en.wikipedia.org/wiki/Graph_mathematics}
}
\keyword{datasets}
